+++ app/src/main/java/com/a01luisrene/multirecordatorio/ui/DetalleRecordatorioActivity.java
@@ -14,17 +14,22 @@ import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
+import android.view.animation.Animation;
+import android.view.animation.AnimationUtils;
 
 import com.a01luisrene.multirecordatorio.MainActivity;
 import com.a01luisrene.multirecordatorio.R;
 import com.a01luisrene.multirecordatorio.fragmentos.ActualizarRecordatorioFragmento;
 import com.a01luisrene.multirecordatorio.fragmentos.AgregarRecordatorioFragmento;
 import com.a01luisrene.multirecordatorio.fragmentos.DetalleRecordatorioFragmento;
+import com.a01luisrene.multirecordatorio.interfaces.InterfaceCrud;
 import com.a01luisrene.multirecordatorio.modelos.Recordatorios;
 import com.a01luisrene.multirecordatorio.sqlite.DataBaseManagerRecordatorios;
 
 
-public class DetalleRecordatorioActivity extends AppCompatActivity implements View.OnClickListener {
+public class DetalleRecordatorioActivity extends AppCompatActivity
+        implements View.OnClickListener,
+        InterfaceCrud {
 
     //Constantes
     public static final int MILISEGUNDOS_ESPERA = 300;
@@ -32,9 +37,12 @@ public class DetalleRecordatorioActivity extends AppCompatActivity implements Vi
 
     //Variables para ocultar el icono eliminar del menú
     public int estadoIconoEliminar;
-    MenuItem eliminarItem;
+    private MenuItem eliminarItem;
 
-    boolean respuestaRetornoEliminarRecordatorio = false;
+    private boolean respuestaRetornoEliminarRecordatorio = false;
+
+    private Animation fab_abrir,fab_cerrar;
+    private Handler handler;
 
     CollapsingToolbarLayout mCtCategoriaRecordatorio;
     FloatingActionButton mFabEditar;
@@ -48,11 +56,14 @@ public class DetalleRecordatorioActivity extends AppCompatActivity implements Vi
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_detalle_recordatorio);
+        //Cierro la actividad si el dispositivo en orientación LANDSCAPE
+        if (getResources().getConfiguration().orientation == Configuration.ORIENTATION_LANDSCAPE) {
+            finish();
+            return;
+        }
 
         mManagerRecordatorios = new DataBaseManagerRecordatorios(this);
-
         mCtCategoriaRecordatorio = (CollapsingToolbarLayout) findViewById(R.id.ct_categoria_recordatorio);
-
         //Toolbar
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar_detail);
         setSupportActionBar(toolbar);
@@ -60,25 +71,21 @@ public class DetalleRecordatorioActivity extends AppCompatActivity implements Vi
             getSupportActionBar().setDisplayHomeAsUpEnabled(true);
 
         mFabEditar = (FloatingActionButton) findViewById(R.id.fab_editar);
+        fab_abrir = AnimationUtils.loadAnimation(getApplicationContext(), R.anim.fab_open);
+        fab_cerrar = AnimationUtils.loadAnimation(getApplicationContext(), R.anim.fab_close);
         mFabEditar.setOnClickListener(this);
-
-        //Cierro la actividad si el dispositivo en orientación LANDSCAPE
-        if (getResources().getConfiguration().orientation == Configuration.ORIENTATION_LANDSCAPE) {
-            finish();
-            return;
-        }
 
         Bundle datos = this.getIntent().getExtras();
         int agregarRecordatorioFragmento = datos.getInt(MainActivity.CARGAR_NUEVO_RECORDATORIO_FRAGMENTO);
 
-
         mItemsRecordatorios = getIntent().getParcelableExtra(DetalleRecordatorioFragmento.KEY_RECORDATORIO);
-
 
         if (savedInstanceState == null) {
             if(agregarRecordatorioFragmento == MainActivity.VALOR_ENVIADO_NUEVO_RECORDATORIO){
 
                 estadoIconoEliminar = 0;
+                //Oculto el FAB editar
+                mFabEditar.setVisibility(View.INVISIBLE);
 
                 mAgregarRecordatorioFragmento = new AgregarRecordatorioFragmento();
 
@@ -87,21 +94,17 @@ public class DetalleRecordatorioActivity extends AppCompatActivity implements Vi
                         .replace(R.id.fl_contenedor_detalle, mAgregarRecordatorioFragmento)
                         .commit();
 
-                //Oculto el botón editar
-                mFabEditar.setVisibility(View.INVISIBLE);
                 mCtCategoriaRecordatorio.setTitle(getString(R.string.agregar_recordatorio));
-
 
             }else{
 
                 estadoIconoEliminar = 1;
 
-                mDetalleRecordatorioFragmento = DetalleRecordatorioFragmento.newInstance(mItemsRecordatorios);
+                mDetalleRecordatorioFragmento = DetalleRecordatorioFragmento.detalleRecordatorio(mItemsRecordatorios);
                 FragmentTransaction ft = getSupportFragmentManager().beginTransaction();
                 ft.replace(R.id.fl_contenedor_detalle, mDetalleRecordatorioFragmento);
                 ft.commit();
 
-                mFabEditar.setVisibility(View.VISIBLE);
             }
         }
 
@@ -111,20 +114,39 @@ public class DetalleRecordatorioActivity extends AppCompatActivity implements Vi
     public void onClick(View v) {
         switch (v.getId()){
             case R.id.fab_editar:
-
                 mActualizarRecordatorioFragmento = ActualizarRecordatorioFragmento.actualizarRecordatorio(mItemsRecordatorios);
                 FragmentTransaction ft = getSupportFragmentManager().beginTransaction();
                 ft.replace(R.id.fl_contenedor_detalle, mActualizarRecordatorioFragmento);
                 ft.addToBackStack(null);
                 ft.commit();
 
+                //Oculto el botón FAB editar
+                animateFAB(fab_cerrar, false);
+                tamanioFabCero(500); //Obliga que el fab scale a 0
 
                 //Oculto el icono de eliminar del menu
                 eliminarItem.setVisible(false);
-
-                //Oculto el botón editar
-                mFabEditar.setVisibility(View.INVISIBLE);
+                break;
         }
+    }
+
+    public void animateFAB(Animation animation, boolean clic){
+        mFabEditar.startAnimation(animation);
+        if(clic){
+            mFabEditar.setClickable(true);
+        }else{
+            mFabEditar.setClickable(false);
+        }
+    }
+    public void tamanioFabCero(int milisegundos){
+        handler = new Handler();
+        handler.postDelayed(new Runnable() {
+            public void run() {
+            // acciones que se ejecutan tras los milisegundos
+            mFabEditar.setScaleX(0);
+            mFabEditar.setScaleY(0);
+            }
+        }, milisegundos);
     }
 
     @Override
@@ -160,50 +182,49 @@ public class DetalleRecordatorioActivity extends AppCompatActivity implements Vi
 
     public void eliminarRecordatorio(){
         new AlertDialog.Builder(this)
-                .setIcon(R.drawable.ic_info_24dp)
-                .setTitle(getString(R.string.adb_titulo_eliminar))
-                .setMessage(getString(R.string.adb_mensaje_eliminar))
-                .setPositiveButton(getString(R.string.boton_aceptar), new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialogInterface, int i) {
+            .setIcon(R.drawable.ic_info_24dp)
+            .setTitle(getString(R.string.adb_titulo_eliminar))
+            .setMessage(getString(R.string.adb_mensaje_eliminar))
+            .setPositiveButton(getString(R.string.boton_aceptar), new DialogInterface.OnClickListener() {
+                @Override
+                public void onClick(DialogInterface dialogInterface, int i) {
 
-                        try {
+                try {
+                    //Id recordatorio
+                    String idRecordatorio = mItemsRecordatorios.getId();
 
-                            //Eliminar registro
-                            String idRecordatorio = mItemsRecordatorios.getId();
-                            mManagerRecordatorios.eliminarRecordatorio(idRecordatorio);
+                    mManagerRecordatorios.eliminarRecordatorio(idRecordatorio);
 
-                            respuestaRetornoEliminarRecordatorio = true;//Respuesta
-                            Intent intent = new Intent();
-                            intent.putExtra(LLAVE_RETORNO_ELIMINAR_RECORDATORIO, respuestaRetornoEliminarRecordatorio);
-                            setResult(RESULT_OK, intent);
+                    respuestaRetornoEliminarRecordatorio = true;//Respuesta
+                    Intent intent = new Intent();
+                    intent.putExtra(LLAVE_RETORNO_ELIMINAR_RECORDATORIO, respuestaRetornoEliminarRecordatorio);
+                    setResult(RESULT_OK, intent);
 
-                            esperarYCerrar(MILISEGUNDOS_ESPERA);
+                    esperarYCerrar(MILISEGUNDOS_ESPERA);
 
-                        } catch (Exception e) {
-                            e.printStackTrace();
-                        }
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
 
-                    }
-                })
-                .setNegativeButton(getString(R.string.boton_cancelar), new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialogInterface, int i) {
-                        //Desplegar mensaje de lamentación
+                }
+            })
+            .setNegativeButton(getString(R.string.boton_cancelar), new DialogInterface.OnClickListener() {
+                @Override
+                public void onClick(DialogInterface dialogInterface, int i) {
+                    //Desplegar mensaje de lamentación
 
-                    }
-                })
-                .show();
+                }
+            })
+            .show();
     }
 
     //Esperar unos segundos antes de cerrar la activity
     public void esperarYCerrar(int milisegundos) {
-        Handler handler = new Handler();
+        handler = new Handler();
         handler.postDelayed(new Runnable() {
             public void run() {
                 // acciones que se ejecutan tras los milisegundos
                 // [cerrar activity]
-
                 finish();
 
             }
@@ -216,7 +237,21 @@ public class DetalleRecordatorioActivity extends AppCompatActivity implements Vi
         //Mostrar icono eliminar del menú
         if (estadoIconoEliminar == 1){
             eliminarItem.setVisible(true);
-            mFabEditar.setVisibility(View.VISIBLE);
         }
+        //Muestro el botón FAB agregar
+        animateFAB(fab_abrir, true);
+        mFabEditar.setScaleX(1);
+        mFabEditar.setScaleY(1);
+    }
+
+
+    @Override
+    public void removerItem() {
+        //Vacío
+    }
+
+    @Override
+    public void actualizarItem(Recordatorios recordatorios) {
+        //Vacío
     }
 }
